django_backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo[srv]==4.6.1
PyJWT==2.8.0
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
bcrypt==4.0.1
